refactor(search): Extract shared ServicePills component
Both ResultItem (list view) and TraceTable (table view) now use the
same ServicePill rendering logic and CSS, including the red-circle
error icon from the list view. ServicePill is an unexported
implementation detail of ServicePills.
