+++ packages/jaeger-ui/src/components/SearchTracePage/SearchResults/ResultItem.css
@@ -13,22 +13,6 @@ SPDX-License-Identifier: Apache-2.0
   border-color: var(--border-strong);
 }
 
-.ResultItem--serviceTag {
-  border-left-width: 15px;
-  margin: 0;
-  display: flex;
-  align-items: center;
-}
-
-.ResultItem--errorIcon {
-  background: var(--feedback-error);
-  border-radius: 6.5px;
-  color: var(--text-inverse);
-  font-size: 0.85em;
-  margin-right: 0.25rem;
-  padding: 1px;
-}
-
 .ResultItem--warningIcon {
   color: var(--feedback-warning);
   font-size: 0.95em;
