+++ packages/jaeger-ui/src/components/SearchTracePage/SearchResults/ServicePills.css
@@ -1,0 +1,20 @@
+/*
+Copyright (c) 2026 The Jaeger Authors.
+SPDX-License-Identifier: Apache-2.0
+*/
+
+.ServicePills--tag {
+  border-left-width: 6px;
+  margin: 0;
+  display: flex;
+  align-items: center;
+}
+
+.ServicePills--errorIcon {
+  background: var(--feedback-error);
+  border-radius: 6.5px;
+  color: var(--text-inverse);
+  font-size: 0.85em;
+  margin-right: 0.25rem;
+  padding: 1px;
+}
